gameservices/blizzard_bounce/requirements.txt
box2d-py==2.3.8
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-sock==0.6.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
Jinja2==3.1.2
simple-websocket==0.9.0
SQLAlchemy==1.4.46
swig==4.2.0
urllib3==1.26.15
Werkzeug==2.2.2
WTForms==3.0.1
